+++ main-app/directivo/asignaturas.php
@@ -46,9 +46,11 @@
 											<div class="row" style="margin-bottom: 10px;">
 												<div class="col-sm-12">
 													<div class="btn-group">
-														<a href="asignaturas-agregar.php" id="addRow" class="btn deepPink-bgcolor">
-															Agregar nuevo <i class="fa fa-plus"></i>
-														</a>
+														<?php if(Modulos::validarPermisoEdicion()){?>
+															<a href="asignaturas-agregar.php" id="addRow" class="btn deepPink-bgcolor">
+																Agregar nuevo <i class="fa fa-plus"></i>
+															</a>
+														<?php }?>
 													</div>
 												</div>
 											</div>
@@ -65,7 +67,9 @@
 														<?php }?>	
 														<th><?=$frases[93][$datosUsuarioActual[8]];?></th>
 														<th>Cargas</th>
-														<th><?=$frases[54][$datosUsuarioActual[8]];?></th>
+														<?php if(Modulos::validarPermisoEdicion()){?>
+															<th><?=$frases[54][$datosUsuarioActual[8]];?></th>
+														<?php }?>
                                                     </tr>
                                                 </thead>
                                                 <tbody>
@@ -98,18 +102,20 @@
 														<td><?=$resultado['ar_nombre'];?></td>
 														<td><a href="cargas.php?asignatura=<?=$resultado['mat_id'];?>" style="text-decoration: underline;"><?=$numeros[0];?></a></td>
 														
-														<td>
-															<div class="btn-group">
-																  <button type="button" class="btn btn-primary"><?=$frases[54][$datosUsuarioActual[8]];?></button>
-																  <button type="button" class="btn btn-primary dropdown-toggle m-r-20" data-toggle="dropdown">
-																	  <i class="fa fa-angle-down"></i>
-																  </button>
-																  <ul class="dropdown-menu" role="menu">
-																	  <li><a href="asignaturas-editar.php?id=<?=$resultado[0];?>"><?=$frases[165][$datosUsuarioActual[8]];?></a></li>
-																	  <?php if($numeros[0]==0){?><li><a href="asignaturas-eliminar.php?id=<?=$resultado[0];?>" onClick="if(!confirm('Desea eliminar este registro?')){return false;}">Eliminar</a></li><?php }?>
-																  </ul>
-															  </div>
-														</td>
+														<?php if(Modulos::validarPermisoEdicion()){?>
+															<td>
+																<div class="btn-group">
+																	<button type="button" class="btn btn-primary"><?=$frases[54][$datosUsuarioActual[8]];?></button>
+																	<button type="button" class="btn btn-primary dropdown-toggle m-r-20" data-toggle="dropdown">
+																		<i class="fa fa-angle-down"></i>
+																	</button>
+																	<ul class="dropdown-menu" role="menu">
+																		<li><a href="asignaturas-editar.php?id=<?=$resultado[0];?>"><?=$frases[165][$datosUsuarioActual[8]];?></a></li>
+																		<?php if($numeros[0]==0){?><li><a href="asignaturas-eliminar.php?id=<?=$resultado[0];?>" onClick="if(!confirm('Desea eliminar este registro?')){return false;}">Eliminar</a></li><?php }?>
+																	</ul>
+																</div>
+															</td>
+														<?php }?>
                                                     </tr>
 													<?php 
 														 $contReg++;
